Post-merge tweaks.
