+++ protocol/bc/output_commitment.go
@@ -5,7 +5,6 @@ import (
 	"io"
 
 	"chain/encoding/blockchain"
-	"chain/encoding/bufpool"
 	"chain/errors"
 )
 
@@ -17,50 +16,49 @@ type OutputCommitment struct {
 	ControlProgram []byte
 }
 
-func (oc *OutputCommitment) writeTo(w io.Writer, assetVersion uint64) (err error) {
-	b := bufpool.Get()
-	defer bufpool.Put(b)
-	if assetVersion == 1 {
-		err = oc.AssetAmount.writeTo(b)
-		if err != nil {
-			return errors.Wrap(err, "writing asset amount")
-		}
-
-		_, err = blockchain.WriteVarint63(b, oc.VMVersion)
-		if err != nil {
-			return errors.Wrap(err, "writing vm version")
-		}
-		_, err = blockchain.WriteVarstr31(b, oc.ControlProgram)
-		if err != nil {
-			return err
-		}
+func (oc *OutputCommitment) WriteTo(w io.Writer) (int64, error) {
+	n, err := oc.AssetAmount.writeTo(w)
+	if err != nil {
+		return n, err
 	}
-
-	_, err = blockchain.WriteVarstr31(w, b.Bytes())
-	return errors.Wrap(err, "writing control program")
+	n2, err := blockchain.WriteVarint63(w, oc.VMVersion)
+	n += int64(n2)
+	if err != nil {
+		return n, err
+	}
+	n2, err = blockchain.WriteVarstr31(w, oc.ControlProgram)
+	n += int64(n2)
+	return n, err
 }
 
-func (oc *OutputCommitment) readFrom(r io.Reader, txVersion, assetVersion uint64) (n int, err error) {
-	if assetVersion != 1 {
-		return n, fmt.Errorf("unrecognized asset version %d", assetVersion)
+// does not write the enclosing extensible string
+func (oc *OutputCommitment) writeWitness(w io.Writer) error {
+	return nil
+}
+
+func (oc *OutputCommitment) ReadFrom(r io.Reader) (int64, error) {
+	n, err := oc.AssetAmount.readFrom(r)
+	if err != nil {
+		return int64(n), errors.Wrap(err, "reading asset+amount")
 	}
-	all := txVersion == 1
-	return blockchain.ReadExtensibleString(r, all, func(r io.Reader) error {
-		_, err := oc.AssetAmount.readFrom(r)
-		if err != nil {
-			return errors.Wrap(err, "reading asset+amount")
-		}
 
-		oc.VMVersion, _, err = blockchain.ReadVarint63(r)
-		if err != nil {
-			return errors.Wrap(err, "reading VM version")
-		}
+	var n2 int
+	oc.VMVersion, n2, err = blockchain.ReadVarint63(r)
+	n += n2
+	if err != nil {
+		return int64(n), errors.Wrap(err, "reading VM version")
+	}
 
-		if oc.VMVersion != 1 {
-			return fmt.Errorf("unrecognized VM version %d for asset version 1", oc.VMVersion)
-		}
+	if oc.VMVersion != 1 {
+		return int64(n), fmt.Errorf("unrecognized VM version %d for asset version 1", oc.VMVersion)
+	}
 
-		oc.ControlProgram, _, err = blockchain.ReadVarstr31(r)
-		return errors.Wrap(err, "reading control program")
-	})
+	oc.ControlProgram, n2, err = blockchain.ReadVarstr31(r)
+	n += n2
+	return int64(n), errors.Wrap(err, "reading control program")
+}
+
+// does not read the enclosing extensible string
+func (oc *OutputCommitment) readWitness(r io.Reader) error {
+	return nil
 }
